+++ NEWS.md
@@ -1,5 +1,7 @@
 # tutorial.helpers (development version)
 
+* Fix format_tutorial() to deal with changes in parsermd 0.1.3.
+
 * Get test cases for format_tutorial() working again.
 
 # tutorial.helpers 0.2.6